refactoring frontend from having static content
