+++ frontend/src/config.json
@@ -1,0 +1,1 @@
+lib/config.json
