+++ frontend/src/lib/config.json
@@ -1,0 +1,5 @@
+{
+  "config": {
+    "dataUrl": "https://raw.githubusercontent.com/joseph-flinn/site/refactor/static-database/data"
+  }
+}
