+++ frontend/src/routes/posts/+page.js
@@ -1,5 +1,6 @@
-import posts from '../../posts.json';
+import { config } from '$lib/config.json';
 
-export const load = () => {
-    return posts;
+export const load = ({fetch}) => {
+    return fetch(`${config.dataUrl}/posts.json`).then(response => response.json());
+
 }
